Use optional authentication
The last release of the kerberos plugin used optional authentication
that was removed as it seemed to be to simplify testing. After further
work it appears this reduces the amount of error messages that can be
received.

When connecting to a kerberos auth endpoint if an error message is
returned on authentication we can't assert the origin of that message
(as we haven't authenticated). Optional means we should accept the body
of these responses for debugging.

Change-Id: I026426cf983c0d7246a559d00f16bd59be804232